warnings (Fabian Greffrath)
https://svn.prboom.org/repos/branches/prboom-plus-24/prboom2/?p=4537

+++ prboom/src/d_deh.c
@@ -1989,25 +1989,25 @@ static void deh_procFrame(DEHFILE *fpin, FILE* fpout, char *line)
         }
       if (!deh_strcasecmp(key,deh_state[0]))  // Sprite number
         {
-          if (fpout) fprintf(fpout," - sprite = %I64d\n",value);
+          if (fpout) fprintf(fpout," - sprite = %ld\n",(long)value);
           states[indexnum].sprite = (spritenum_t)value;
         }
       else
         if (!deh_strcasecmp(key,deh_state[1]))  // Sprite subnumber
           {
-            if (fpout) fprintf(fpout," - frame = %I64d\n",value);
+            if (fpout) fprintf(fpout," - frame = %ld\n",(long)value);
             states[indexnum].frame = (long)value; // long
           }
         else
           if (!deh_strcasecmp(key,deh_state[2]))  // Duration
             {
-              if (fpout) fprintf(fpout," - tics = %I64d\n",value);
+              if (fpout) fprintf(fpout," - tics = %ld\n",(long)value);
               states[indexnum].tics = (long)value; // long
             }
           else
             if (!deh_strcasecmp(key,deh_state[3]))  // Next frame
               {
-                if (fpout) fprintf(fpout," - nextstate = %I64d\n",value);
+                if (fpout) fprintf(fpout," - nextstate = %ld\n",(long)value);
                 states[indexnum].nextstate = (statenum_t)value;
               }
             else
@@ -2019,13 +2019,13 @@ static void deh_procFrame(DEHFILE *fpin, FILE* fpout, char *line)
               else
                 if (!deh_strcasecmp(key,deh_state[5]))  // Unknown 1
                   {
-                    if (fpout) fprintf(fpout," - misc1 = %I64d\n",value);
+                    if (fpout) fprintf(fpout," - misc1 = %ld\n",(long)value);
                     states[indexnum].misc1 = (long)value; // long
                   }
                 else
                   if (!deh_strcasecmp(key,deh_state[6]))  // Unknown 2
                     {
-                      if (fpout) fprintf(fpout," - misc2 = %I64d\n",value);
+                      if (fpout) fprintf(fpout," - misc2 = %ld\n",(long)value);
                       states[indexnum].misc2 = (long)value; // long
                     }
                   else
@@ -2082,15 +2082,15 @@ static void deh_procPointer(DEHFILE *fpin, FILE* fpout, char *line) // done
       if (value >= NUMSTATES)
         {
           if (fpout)
-            fprintf(fpout,"Bad pointer number %I64d of %d\n",value, NUMSTATES);
+            fprintf(fpout,"Bad pointer number %ld of %d\n",(long)value, NUMSTATES);
           return;
         }
 
       if (!deh_strcasecmp(key,deh_state[4]))  // Codep frame (not set in Frame deh block)
         {
           states[indexnum].action = deh_codeptr[value];
-          if (fpout) fprintf(fpout," - applied from codeptr[%I64d] to states[%d]\n",
-           value,indexnum);
+          if (fpout) fprintf(fpout," - applied from codeptr[%ld] to states[%d]\n",
+           (long)value,indexnum);
           // Write BEX-oriented line to match:
           // for (i=0;i<NUMSTATES;i++) could go past the end of the array
           for (i=0;i<sizeof(deh_bexptrs)/sizeof(*deh_bexptrs);i++)
@@ -2106,8 +2106,8 @@ static void deh_procPointer(DEHFILE *fpin, FILE* fpout, char *line) // done
             }
         }
       else
-        if (fpout) fprintf(fpout,"Invalid frame pointer index for '%s' at %I64d\n",
-                           key, value);
+        if (fpout) fprintf(fpout,"Invalid frame pointer index for '%s' at %ld\n",
+                           key, (long)value);
     }
   return;
 }
@@ -2792,7 +2792,7 @@ static void deh_procStrings(DEHFILE *fpin, FILE* fpout, char *line)
     // killough 11/98: allocate enough the first time
           maxstrlen = strlen(holdstring) + strlen(inbuffer);
           if (fpout) fprintf(fpout,
-                             "* increased buffer from to %d for buffer size %d\n",
+                             "* increased buffer from to %ld for buffer size %d\n",
                              maxstrlen,(int)strlen(inbuffer));
           holdstring = realloc(holdstring,maxstrlen*sizeof(*holdstring));
         }
